:sparkles: feat: 사용자 고유 데이터 인가 권한 설정

+++ backend/src/main/java/com/s12p31b204/backend/controller/TestPaperController.java
@@ -5,6 +5,7 @@ import java.util.NoSuchElementException;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,6 +18,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.s12p31b204.backend.dto.CreateTestPaperRequestDto;
 import com.s12p31b204.backend.dto.TestPaperResponseDto;
 import com.s12p31b204.backend.dto.UpdateTestPaperRequestDto;
+import com.s12p31b204.backend.oauth2.CustomOAuth2User;
+import com.s12p31b204.backend.service.AuthorizationService;
 import com.s12p31b204.backend.service.TestPaperService;
 import com.s12p31b204.backend.util.ApiResponse;
 import com.s12p31b204.backend.util.ResponseData;
@@ -32,13 +35,21 @@ import lombok.extern.slf4j.Slf4j;
 public class TestPaperController {
 
     private final TestPaperService testPaperService;
+    private final AuthorizationService authorizationService;
 
     @GetMapping("/{workBookId}")
-    public ResponseEntity<ResponseData<List<TestPaperResponseDto>>> getTestPapers(@PathVariable Long workBookId, HttpServletRequest request) {
+    public ResponseEntity<ResponseData<List<TestPaperResponseDto>>> getTestPapers
+            (@PathVariable Long workBookId,
+             @AuthenticationPrincipal CustomOAuth2User user,
+             HttpServletRequest request) {
         try {
-            log.info("getting TestPapers...");
-            List<TestPaperResponseDto> testPapers = testPaperService.findTestPaperByWorkBookId(workBookId);
-            return ApiResponse.success(testPapers, "시험지 리스트 조회 성공", HttpStatus.OK, request.getRequestURI());
+            if(authorizationService.checkWorkBookAuthorization(user.getUserId(), workBookId)) {
+                log.info("getting TestPapers...");
+                List<TestPaperResponseDto> testPapers = testPaperService.findTestPaperByWorkBookId(workBookId);
+                return ApiResponse.success(testPapers, "시험지 리스트 조회 성공", HttpStatus.OK, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (NoSuchElementException e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.BAD_REQUEST, request.getRequestURI());
         } catch (Exception e) {
@@ -48,11 +59,19 @@ public class TestPaperController {
     }
 
     @PostMapping
-    public ResponseEntity<ResponseData<TestPaperResponseDto>> createTestPaper(@RequestBody CreateTestPaperRequestDto createTestPaperRequestDto, HttpServletRequest request) {
+    public ResponseEntity<ResponseData<TestPaperResponseDto>> createTestPaper(
+            @RequestBody CreateTestPaperRequestDto createTestPaperRequestDto,
+            @AuthenticationPrincipal CustomOAuth2User user,
+            HttpServletRequest request
+    ) {
         try {
-            log.info("creating TestPaper...");
-            TestPaperResponseDto response = testPaperService.createTestPaper(createTestPaperRequestDto);
-            return ApiResponse.success(response, "시험지 생성 성공", HttpStatus.CREATED, request.getRequestURI());
+            if(authorizationService.checkWorkBookAuthorization(user.getUserId(), createTestPaperRequestDto.getWorkBookId())) {
+                log.info("creating TestPaper...");
+                TestPaperResponseDto response = testPaperService.createTestPaper(createTestPaperRequestDto);
+                return ApiResponse.success(response, "시험지 생성 성공", HttpStatus.CREATED, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (NoSuchElementException e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.BAD_REQUEST, request.getRequestURI());
         } catch (Exception e) {
@@ -62,11 +81,18 @@ public class TestPaperController {
     }
 
     @DeleteMapping("/{testPaperId}")
-    public ResponseEntity<ResponseData<Void>> removeTestPaper(@PathVariable Long testPaperId, HttpServletRequest request) {
+    public ResponseEntity<ResponseData<Void>> removeTestPaper(
+            @PathVariable Long testPaperId,
+            @AuthenticationPrincipal CustomOAuth2User user,
+            HttpServletRequest request) {
         try {
-            log.info("removing TestPaper...");
-            testPaperService.removeTestPaper(testPaperId);
-            return ApiResponse.success(null, "시험지 삭제 성공", HttpStatus.NO_CONTENT, request.getRequestURI());
+            if(authorizationService.checkTestPaperAuthorization(user.getUserId(), testPaperId)) {
+                log.info("removing TestPaper...");
+                testPaperService.removeTestPaper(testPaperId);
+                return ApiResponse.success(null, "시험지 삭제 성공", HttpStatus.NO_CONTENT, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (Exception e) {
             log.error(e.getMessage());
             return ApiResponse.failure("시험지 삭제 중 오류 발생", HttpStatus.INTERNAL_SERVER_ERROR, request.getRequestURI());
@@ -74,11 +100,18 @@ public class TestPaperController {
     }
 
     @PutMapping
-    public ResponseEntity<ResponseData<TestPaperResponseDto>> modifyTestPaper(@RequestBody UpdateTestPaperRequestDto updateTestPaperRequestDto, HttpServletRequest request) {
+    public ResponseEntity<ResponseData<TestPaperResponseDto>> modifyTestPaper(
+            @RequestBody UpdateTestPaperRequestDto updateTestPaperRequestDto,
+            @AuthenticationPrincipal CustomOAuth2User user,
+            HttpServletRequest request) {
         try {
-            log.info("modifying TestPaper...");
-            TestPaperResponseDto response = testPaperService.updateTestPaper(updateTestPaperRequestDto);
-            return ApiResponse.success(response, "시험지 수정 성공", HttpStatus.OK, request.getRequestURI());
+            if(authorizationService.checkTestPaperAuthorization(user.getUserId(), updateTestPaperRequestDto.getTestPaperId())) {
+                log.info("modifying TestPaper...");
+                TestPaperResponseDto response = testPaperService.updateTestPaper(updateTestPaperRequestDto);
+                return ApiResponse.success(response, "시험지 수정 성공", HttpStatus.OK, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (NoSuchElementException e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.BAD_REQUEST, request.getRequestURI());
         } catch (Exception e) {
